Remove packets waiting for an ARP request after sending them

+++ acl-table.cpp
@@ -55,7 +55,7 @@ namespace simple_router {
 //////////////////////////////////////////////////////////////////////////
 // IMPLEMENT THESE METHODS
 ACLTableEntry
-ACLTable::lookup(uint32_t srcIp, uint32_t dstIp, uint8_t protocol, uint32_t srcPort, uint16_t dstPort) const
+ACLTable::lookup(uint32_t srcIp, uint32_t dstIp, uint8_t protocol, uint16_t srcPort, uint16_t dstPort) const
 {
   // FILL THIS IN
   // Iterate over all entries
