+++ acl-table.hpp
@@ -57,7 +57,7 @@ public:
    * If ACL table not found, `throw std::runtime_error("ACL entry not found")`
    */
   ACLTableEntry
-  lookup(uint32_t srcIp, uint32_t dstIp, uint8_t protocol, uint32_t srcPort, uint16_t dstPort) const;
+  lookup(uint32_t srcIp, uint32_t dstIp, uint8_t protocol, uint16_t srcPort, uint16_t dstPort) const;
 
   bool
   load(const std::string& file);
